requirements.txt
requests~=2.32.3
pydantic~=2.9.2
python-dotenv~=1.0.1
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-core==0.3.11
rich==13.9.2
langgraph==0.2.38